Remove navigate.
This was used for edge cases where a url may stay the same but other factors may cause a navigation to happen. Routr should not be taking this into account since it is simply a route matching library. If needed, the navigation logic should happen at a higher level or remove that special case.