Fix encoding/decoding inconsistency when using values that need scaping
Since `makePath` uses pathToRegexp which in place uses encodeURIComponent for all
values, when getRoute tries to match a URL generated by the router itself, it returns
a double enconded value.

I suspect this is the root cause for fluxible-router to avoid using parsed URLs
and [bookeeping the origURL instead](1)

I added tests to both `match` (indirectly through `getRoute`) and to `makePath`. Notice
that if you comment or remove `decodeURIComponent` from the implementation, only 1 of the
tests fails. This proves the undesired inconsistency.

[1]: https://github.com/yahoo/fluxible/blob/2be37f6df6ef5d5ed7c31fae5ec14843f924b213/packages/fluxible-router/lib/History.js#L66-L73