Only append query if there is something to append
This is mostly cosmetic, but it is really strange to have a question
mark if you don't have query, and adding checks on any application
using routr to not pass a query seems wasteful.

i.e. in my application I use:

`Object.assing({}, detaultQuery, myFuncQueryParam)` and sometimes I
get an empty object. It would be convenient if the router didn't add
the question mark.

    r.makePath('my', {wat:'route'}, {});

Before this patch: /my/route?
After this patch: /my/route